feat: move script editor to reusable component
