+++ frontend/src/app/shared/components/script-editor/script-editor.component.html
@@ -1,0 +1,49 @@
+<div class="flex flex-row gap-2">
+    <div class="w-2/4">
+        <div class="card flex flex-col gap-4">
+            <div class="font-semibold text-xl">Script (lua)</div>
+
+            <ngx-monaco-editor class="min-h-96" (onInit)="onEditorInit($event)" [options]="editorOptions"
+                               [(ngModel)]="script"></ngx-monaco-editor>
+
+        </div>
+    </div>
+    <div class="w-2/4">
+        <div class="card flex flex-col gap-4 h-full">
+            <div class="font-semibold text-xl">Documentation</div>
+            <textarea class="h-full" [ngModel]="helpContent" pTextarea readonly></textarea>
+        </div>
+    </div>
+</div>
+<div class="flex flex-row gap-4">
+    <div class="flex flex-col card gap-4 w-full">
+        <div class="font-semibold text-xl">Dry run</div>
+        <div class="flex flex-row gap-4">
+            <div class="flex flex-col gap-2 w-1/2">
+                <div class="flex flex-row gap-2">
+                    <div class="flex flex-col w-full gap-2">
+                        <label for="accountName">Transaction ID</label>
+                        <input  pInputText id="accountName" type="number" [(ngModel)]="dryRunTransactionId"/>
+                    </div>
+                    <div class="flex flex-col w-full gap-2">
+                        <label for="accountName">Fetch test transaction</label>
+                        <p-button severity="info" (onClick)="ensureTxSet()">Fetch last transaction</p-button>
+                    </div>
+                </div>
+            </div>
+
+            <div class="flex flex-col gap-2 w-1/2">
+
+                <label for="accountName">Dry run</label>
+                <p-button label="Test script" (onClick)="dryRun()" />
+            </div>
+        </div>
+        <div class="flex flex-row">
+            <div class="flex flex-col gap-4 w-full">
+                <div class="font-semibold text-xl">Diff ( Original -> Modified )</div>
+
+                <ngx-monaco-diff-editor class="min-h-96" (onInit)="onDiffEditorInit($event)" [options]="diffOptions" [originalModel]="originalModel" [modifiedModel]="modifiedModel"></ngx-monaco-diff-editor>
+            </div>
+        </div>
+    </div>
+</div>
